+++ pyquil/api/_quantum_computer.py
@@ -17,7 +17,7 @@ import re
 import socket
 import warnings
 from math import pi, log
-from typing import List, Dict, Tuple, Iterator, Union
+from typing import List, Dict, Tuple, Iterator, Mapping, Optional, Sequence, Union
 import itertools
 
 import subprocess
@@ -36,6 +36,8 @@ from pyquil.api._qam import QAM
 from pyquil.api._qpu import QPU
 from pyquil.api._qvm import ForestConnection, QVM
 from pyquil.device import AbstractDevice, NxDevice, gates_in_isa, ISA, Device
+from pyquil.experiment import (ExperimentResult, TomographyExperiment, bitstrings_to_expectations,
+                               merge_memory_map_lists)
 from pyquil.gates import RX, MEASURE
 from pyquil.noise import decoherence_noise_with_asymmetric_ro, NoiseModel
 from pyquil.pyqvm import PyQVM
@@ -127,6 +129,115 @@ class QuantumComputer:
         return self.qam.run() \
             .wait() \
             .read_memory(region_name='ro')
+
+    @_record_call
+    def experiment(
+            self,
+            experiment: TomographyExperiment,
+            memory_map: Optional[Mapping[str, Sequence[Union[int, float]]]] = None
+    ) -> List[ExperimentResult]:
+        """
+        Run a ``TomographyExperiment`` on a QVM or QPU backend. A ``TomographyExperiment``
+        is composed of:
+
+            - A main ``Program`` body (or ansatz).
+            - A collection of ``ExperimentSetting`` objects, each of which encodes a particular
+              state preparation and measurement.
+            - A ``SymmetrizationLevel`` for enacting different readout symmetrization strategies.
+            - A number of shots to collect for each (unsymmetrized) ``ExperimentSetting``.
+
+        Because the main ``Program`` is static from run to run of a ``TomographyExperiment``, we
+        can leverage our platform's Parametric Compilation feature. This means that the ``Program``
+        can be compiled only once, and the various alterations due to state preparation,
+        measurement, and symmetrization can all be realized at runtime by providing a
+        ``memory_map``. Thus, the steps in the ``experiment`` method are as follows:
+
+            1. Check to see if this ``TomographyExperiment`` has already been loaded into this
+               ``QuantumComputer`` object. If so, skip to step 2. Otherwise, do the following:
+
+                a. Generate a parameterized program corresponding to the ``TomographyExperiment``
+                   (see the ``TomographyExperiment.generate_experiment_program()`` method for more
+                   details on how it changes the main body program to support state preparation,
+                   measurement, and symmetrization).
+                b. Compile the parameterized program into a parametric (binary) executable, which
+                   contains declared variables that can be assigned at runtime.
+
+            2. For each ``ExperimentSetting`` in the ``TomographyExperiment``, we repeat the
+               following:
+
+                a. Build a collection of memory maps that correspond to the various state
+                   preparation, measurement, and symmetrization specifications.
+                b. Run the parametric executable on the QVM or QPU backend, providing the memory map
+                   to assign variables at runtime.
+                c. Extract the desired statistics from the classified bitstrings that are produced
+                   by the QVM or QPU backend, and package them in an ``ExperimentResult`` object.
+
+            3. Return the list of ``ExperimentResult`` objects.
+
+        This method is extremely useful shorthand for running near-term applications and algorithms,
+        which often have this ansatz + settings structure.
+
+        :param experiment: The ``TomographyExperiment`` to run.
+        :param memory_map: A dictionary mapping declared variables / parameters to their values.
+            The values are a list of floats or integers. Each float or integer corresponds to
+            a particular classical memory register. The memory map provided to the ``experiment``
+            method corresponds to variables in the main body program that we would like to change
+            at runtime (e.g. the variational parameters provided to the ansatz of the variational
+            quantum eigensolver).
+        :return: A list of ``ExperimentResult`` objects containing the statistics gathered
+            according to the specifications of the ``TomographyExperiment``.
+        """
+        executable = self.qam._executable
+        # if this experiment was the last experiment run on this QuantumComputer,
+        # then use the executable that is already loaded into the object
+        if self.qam._experiment != experiment:
+            experiment_program = experiment.generate_experiment_program()
+            executable = self.compile(experiment_program)
+            self.qam._experiment = experiment
+
+        if memory_map is None:
+            memory_map = {}
+
+        results = []
+        for settings in experiment:
+            # TODO: add support for grouped ExperimentSettings
+            if len(settings) > 1:
+                raise ValueError('We only support length-1 settings for now.')
+            setting = settings[0]
+
+            qubits = setting.out_operator.get_qubits()
+            experiment_setting_memory_map = experiment.build_setting_memory_map(setting)
+            symmetrization_memory_maps = experiment.build_symmetrization_memory_maps(qubits)
+            merged_memory_maps = merge_memory_map_lists([experiment_setting_memory_map],
+                                                        symmetrization_memory_maps)
+
+            all_bitstrings = []
+            # TODO: accomplish symmetrization via batch endpoint
+            for merged_memory_map in merged_memory_maps:
+                final_memory_map = {**memory_map, **merged_memory_map}
+                bitstrings = self.run(executable, memory_map=final_memory_map)
+
+                if 'symmetrization' in final_memory_map:
+                    bitmask = np.array(np.array(final_memory_map['symmetrization']) / np.pi,
+                                       dtype=int)
+                    bitstrings = np.bitwise_xor(bitstrings, bitmask)
+                all_bitstrings.append(bitstrings)
+            symmetrized_bitstrings = np.concatenate(all_bitstrings)
+
+            # TODO: support simultaneous observables via multiple correlations
+            joint_expectations = [experiment.get_meas_registers(qubits)]
+            expectations = bitstrings_to_expectations(symmetrized_bitstrings,
+                                                      joint_expectations=joint_expectations)
+
+            # TODO: add calibration and correction
+            mean = np.mean(expectations).item()
+            std_err = (np.std(expectations, axis=0, ddof=1) / np.sqrt(len(expectations))).item()
+            result = ExperimentResult(setting=setting,
+                                      expectation=mean,
+                                      std_err=std_err,
+                                      total_counts=len(expectations))
+            results.append(result)
+        return results
 
     @_record_call
     def run_symmetrized_readout(self, program: Program, trials: int, symm_type: int = 3,