+++ pyquil/experiment/__init__.py
@@ -1,6 +1,7 @@
-from pyquil.experiment._main import (OperatorEncoder, SymmetrizationLevel, TomographyExperiment,
-                                     read_json, to_json)
-from pyquil.experiment._result import ExperimentResult
+from pyquil.experiment._main import OperatorEncoder, TomographyExperiment, read_json, to_json
+from pyquil.experiment._memory import merge_memory_map_lists
+from pyquil.experiment._result import ExperimentResult, bitstrings_to_expectations
 from pyquil.experiment._setting import (_OneQState, _pauli_to_product_state, ExperimentSetting,
                                         SIC0, SIC1, SIC2, SIC3, TensorProductState, minusX, minusY,
                                         minusZ, plusX, plusY, plusZ, zeros_state)
+from pyquil.experiment._symmetrization import SymmetrizationLevel