+++ pyquil/experiment/_setting.py
@@ -24,6 +24,8 @@ import warnings
 from typing import Iterable, Tuple
 
 from pyquil.paulis import PauliTerm, sI, is_identity
+from pyquil.experiment._memory import (pauli_term_to_measurement_memory_map,
+                                       pauli_term_to_preparation_memory_map)
 
 if sys.version_info < (3, 7):
     from pyquil.external.dataclasses import dataclass