+++ CHANGELOG.md
@@ -8,6 +8,16 @@ Changelog
 
 ### Improvements and Changes
 
+-   There is a new `QuantumComputer.experiment` method for running a collection of
+    quantum programs as defined by a `TomographyExperiment`. These objects have a
+    main program body and a collection of state preparation and measurement
+    specifications, which capture the structure of many near-term applications
+    and algorithms like the variational quantum eigensolver (VQE). In addition,
+    the `TomographyExperiment` encodes information about symmetrization, active
+    qubit reset, and the number of shots to perform on the quantum backend (e.g.
+    the QVM or QPU). For more information check out the API documentation sections
+    on the [Quantum Computer](docs/source/apidocs/quantum_computer.rst) and on the
+    [Experiment Module](docs/source/apidocs/experiment.rst) (@karalekas, gh-1100).
 -   Type hints have been added to the `PauliTerm` class (@rht, gh-1075).
 -   The `rigetti/forest` Docker image now has less noisy output due to stdout and
     stderr redirection to log files `entrypoint.sh` (@karalekas, gh-1105).